services/api/requirements.txt
fastapi==0.128.7
uvicorn[standard]==0.40.0
python-multipart==0.0.22
requests==2.32.5
redis==7.1.1
rq==2.6.1
reportlab==4.4.9
PyYAML==6.0.3
pdfplumber==0.11.9
xlrd==2.0.2
mem0ai==1.0.3
matplotlib==3.10.8
numpy==2.4.2
pandas==3.0.0
seaborn==0.13.2
deepseek-ocr>=0.3.0; python_version>="3.10"
multi-ocr-sdk>=0.5.1; python_version>="3.10"
